+++ src/Sarif.Sarifer/AnalyzeProjectCommand.cs
@@ -1,76 +1,26 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections.Generic;
 using System.ComponentModel.Design;
 using System.Linq;
-using System.Threading;
 
 using EnvDTE;
 
-using EnvDTE80;
-
-using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Shell;
 
 namespace Microsoft.CodeAnalysis.Sarif.Sarifer
 {
-    internal class AnalyzeProjectCommand : IDisposable
+    internal class AnalyzeProjectCommand : AnalyzeMenuCommandBase
     {
-        private DTE2 dte;
-        private IComponentModel componentModel;
-        private IBackgroundAnalysisService backgroundAnalysisService;
-        private readonly IMenuCommandService menuCommandService;
-        private CancellationTokenSource cancellationTokenSource;
-        private bool disposed;
-
         public AnalyzeProjectCommand(IMenuCommandService menuCommandService)
+            : base(menuCommandService, SariferPackageCommandIds.AnalyzeProject)
         {
-            var menuCommand = new MenuCommand(
-                new EventHandler(this.MenuCommandCallback),
-                new CommandID(Guids.SariferCommandSet, SariferPackageCommandIds.AnalyzeProject));
-
-            menuCommandService.AddCommand(menuCommand);
-            this.menuCommandService = menuCommandService;
         }
 
-        /// <summary>
-        /// Cancel current analysis.
-        /// </summary>
-        public void Cancel()
-        {
-            this.backgroundAnalysisService?.ClearResultsAsync()
-                .FileAndForget(FileAndForgetEventName.BackgroundAnalysisFailure);
-            this.cancellationTokenSource?.Cancel();
-        }
-
-        /// <summary>
-        /// Event handler called when the user selects the Analyze Project command.
-        /// </summary>
-        private void MenuCommandCallback(object caller, EventArgs args)
+        protected override void AnalyzeTargets()
         {
             ThreadHelper.ThrowIfNotOnUIThread();
-
-            // Always cancel before start.
-            this.cancellationTokenSource?.Cancel();
-            this.cancellationTokenSource = new CancellationTokenSource();
-
-            if (this.dte == null)
-            {
-                this.dte = (DTE2)Package.GetGlobalService(typeof(DTE));
-            }
-
-            if (this.componentModel == null)
-            {
-                this.componentModel = (IComponentModel)Package.GetGlobalService(typeof(SComponentModel));
-            }
-
-            if (this.backgroundAnalysisService == null)
-            {
-                this.backgroundAnalysisService = this.componentModel.GetService<IBackgroundAnalysisService>();
-                this.backgroundAnalysisService.AnalysisCompleted += this.BackgroundAnalysisService_AnalysisCompleted;
-            }
 
             IEnumerable<Project> selectedProjects = (this.dte.ActiveSolutionProjects as object[]).OfType<Project>();
             if (selectedProjects != null)
@@ -86,32 +36,6 @@ namespace Microsoft.CodeAnalysis.Sarif.Sarifer
                         .FileAndForget(FileAndForgetEventName.BackgroundAnalysisFailure);
                 }
             }
-        }
-
-        private void BackgroundAnalysisService_AnalysisCompleted(object sender, EventArgs e)
-        {
-            SariferPackageCommand.EnableAnalyzeCommands(this.menuCommandService);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!this.disposed)
-            {
-                if (disposing)
-                {
-                    this.cancellationTokenSource?.Cancel();
-                    this.cancellationTokenSource?.Dispose();
-                    this.backgroundAnalysisService.AnalysisCompleted -= this.BackgroundAnalysisService_AnalysisCompleted;
-                }
-
-                this.disposed = true;
-            }
-        }
-
-        public void Dispose()
-        {
-            this.Dispose(disposing: true);
-            GC.SuppressFinalize(this);
         }
     }
 }
